Resolving dominator cycle using cache for seen nodes (#9) (WIP)
Changed the map to use hash instead of node pointers because the pointers for same node change under different contexts.
Using a seen cache resolves the dominator cycle but the results still need improvement. History for intermediate nodes is not dumped correctly.
With the map key change, a code overhaul is needed to get rid of now useless code + some other repeated code.