Added TypeSummary & fixed HistoryContext (WIP) #14
TypeSummary class is intended to be copyable and composable (truly)
to record history of cast operand types. Currently dominator cycles
are not supported.

Keeping entries in HistoryContext that resolve a parameter with another
parameter, is not very useful and is usually skipped to avoid spurious
resolutions using the context. Now such resolutions are not added to
context during context construction.