LocalHistory now holds history branch info (WIP #22).
Earlier domdata was only available in census which makes it difficult to
analyze the cast type in type summary, esp. for function calls.

It is now a part of localhistory which is now an improved
abstraction for type history graph node. LocalHistory stores context as
well as the metadata for the link from one History to the other.
In case of the following cast:
    int *pi = (int*) pv;

History(pv) = void ---[domdata(pv->pi)]---> History(pi)
                      ^---------------^

The linkdata is stored through LocalHistory and can store the type of
cast involved for this link (BitCast, LToR, etc.).